# app/Models/Location.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Concerns\HasUuids;
use Illuminate\Database\Eloquent\Model;

class Location extends Model
{

	use HasUuids;

	public $timestamps = false;

	protected $fillable = [
		'name',
		'code',
	];
}